+++ scratch.js
@@ -1,4 +1,5 @@
 const spacetime = require('./src/index')
 
 let s = spacetime('March 2')
-console.log(s.offset())
+console.log(s.format('dd'))
+console.log(s.unixFmt('dd-MMM'))