+++ changelog.md
@@ -72,3 +72,5 @@
  - - existing code will now need to be `s = s.add(1,'day')`
  - - - (instead of `s.add(1,'day')`)
 * uppercase month and day names in `.dayName()`, etc.
+* move unix-formating out of `.format()` into `.unixFmt()`
+* change default `.format()` response to **yyyy-mm-dd**